Extending asset type with external data
